pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.2",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
